Reorganized how dbConnections are imported and used
