+++ Service/src/nuke.py
@@ -1,7 +1,6 @@
 from flask import Blueprint, request, Response, g
 from util import logger
 import pymysql
-import dbConn
 
 log = logger.Logger(True, True, True)
