+++ Service/src/transaction.py
@@ -3,7 +3,6 @@ from util import logger
 from datetime import datetime
 import pymysql
 import simplejson as json
-import dbConn
 
 log = logger.Logger(True, True, True)
