+++ Service/src/user.py
@@ -2,7 +2,6 @@ from flask import Blueprint, request, Response, g
 from util import logger
 from werkzeug.security import generate_password_hash, check_password_hash
 import pymysql
-import dbConn
 
 log = logger.Logger(True, True, True)
